requirements.txt
tensorflow-gpu==2.0.0
numpy==1.17.0
pandas==1.0.3
opencv-python==4.2.0.32
matplotlib==3.1.0
cloudpickle==1.3.0
scikit-learn==0.22.2
pillow==7.0.0
